+++ api/services/PUT.js
@@ -1,4 +1,4 @@
-import Incidents from "../collections/incidents.js";
+import Incidents from "../models/incidents.js";
 
 const incident = new Incidents();
